+++ external_agent/examples/langgraph_python/tools.py
@@ -1,0 +1,21 @@
+from langchain_core.tools import tool
+from langchain_community.tools import DuckDuckGoSearchResults
+
+@tool
+def web_search_duckduckgo(search_phrase: str):
+    """Search the web using duckduckgo."""
+    search = DuckDuckGoSearchResults()
+    results = search.run(search_phrase) 
+    return results
+
+@tool
+def news_search_duckduckgo(search_phrase: str):
+    """Search news using duckduckgo."""
+    search = DuckDuckGoSearchResults(backend="news")
+    results = search.run(search_phrase) 
+    return results
+
+tool_choices = {
+    "web_search_duckduckgo": web_search_duckduckgo,
+    "news_search_duckduckgo": news_search_duckduckgo,
+}
